test: fix broken cli tests and start running them on validate (#57)
* feat(tests): run cli-tests in CI

In order to make sure the prettier-eslint command won't display colored
output, we have to tweak the environment variables. We don't want
colored output because it won't match jest snapshots.

* fix(tests): adjust expected output for stdout cli tests

In #54 I removed the extra new line at the end of files printed to
stdout. I wasn't aware of cli-tests, so I didn't change those.

This commit adjust the expected output of files formated and printed to
stdout.