Removed default values for eslintPath and prettierPath
These values are not required by prettier-eslint-cli itself. They
are only passed to prettier-eslint, and even then, they're optional.
When using npm >3, passing the default value is optional, since
the node_modules directory is flattened, but for npm 2.x, the
default assumptions don't work.

People should be running npm >3 today, but for CI scenarios, npm 2
or downwards can still be used in combination with node 4.x (which
is going to be supported for a long time still).

In any case, not passing a default value seems to align better with
node's require algorithm - it will start where it needs to start
(in this case wherever the prettier-eslint dependency has been
installed), and then mimick find-up's logic.